// client/src/app/core/api/models/Tale.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Readable } from './Readable';
export type Tale = {
    id?: string;
    tale?: Array<Readable>;
    author?: string;
    published?: boolean;
    writeDate?: string;
    publishDate?: string | null;
};